[sw/runtime] move the irq lib to sw/device/lib/runtime
This moves the irq runtime library to the same location as other similar
runtime libraries, like the epmp and ibex libs.
