+++ sw/device/lib/testing/test_framework/BUILD
@@ -239,10 +239,10 @@ cc_library(
         ":freertos_config",
         ":ottf_start",
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
-        "//sw/device/lib:irq",
         "//sw/device/lib/dif:rv_timer",
         "//sw/device/lib/dif:uart",
         "//sw/device/lib/runtime:hart",
+        "//sw/device/lib/runtime:irq",
         "//sw/device/lib/runtime:log",
         "//third_party/freertos",
     ],
