+++ sw/device/lib/testing/test_framework/freertos_hooks.c
@@ -4,8 +4,8 @@
 
 #include "external/freertos/include/FreeRTOS.h"
 #include "external/freertos/include/task.h"
-#include "sw/device/lib/irq.h"
 #include "sw/device/lib/runtime/hart.h"
+#include "sw/device/lib/runtime/irq.h"
 #include "sw/device/lib/runtime/log.h"
 
 // NOTE: the function names below do NOT, and cannot, conform to the style
